# Checkpoint of the world state should be moved by FCU