--- a/src/Nethermind/Nethermind.Blockchain.Test/BlockchainProcessorTests.cs
+++ b/src/Nethermind/Nethermind.Blockchain.Test/BlockchainProcessorTests.cs
@@ -113,6 +113,10 @@ public class BlockchainProcessorTests
                 }
             }
 
+            public void UpdateCheckpoint(long blockNumber, Hash256 stateRootCheckpoint)
+            {
+            }
+
             public event EventHandler<BlocksProcessingEventArgs>? BlocksProcessing;
 
             public event EventHandler<BlockEventArgs>? BlockProcessing;