--- a/src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.cs
+++ b/src/Nethermind/Nethermind.Consensus/Processing/BlockProcessor.cs
@@ -276,7 +276,7 @@ public partial class BlockProcessor(
 
     public void UpdateCheckpoint(long blockNumber, Hash256 stateRootCheckpoint)
     {
-        if (_logger.IsInfo) _logger.Info($"Updating checkpoint to {stateRootCheckpoint} at block {blockNumber}");
+        if (_logger.IsTrace) _logger.Trace($"Updating checkpoint to {stateRootCheckpoint} at block {blockNumber}");
         _checkpoint = stateRootCheckpoint;
     }
 
@@ -290,10 +290,10 @@ public partial class BlockProcessor(
     // TODO: move to branch processor
     private void RestoreBranch(Hash256 branchingPointStateRoot)
     {
-        if (_logger.IsInfo) _logger.Info($"Restoring the branch checkpoint - {branchingPointStateRoot}");
+        if (_logger.IsTrace) _logger.Trace($"Restoring the branch checkpoint - {branchingPointStateRoot}");
         _stateProvider.Reset();
         _stateProvider.StateRoot = branchingPointStateRoot;
-        if (_logger.IsInfo) _logger.Info($"Restored the branch checkpoint - {branchingPointStateRoot} | {_stateProvider.StateRoot}");
+        if (_logger.IsTrace) _logger.Trace($"Restored the branch checkpoint - {branchingPointStateRoot} | {_stateProvider.StateRoot}");
     }
 
     // TODO: block processor pipeline