@LukaszRozmej 2025-01-31
I think it is all not needed. What is really needed is in the ForkChoiceUpdate handler move main state provider state root to the root of the new head block