@LukaszRozmej 2025-02-03
Alternative I had in mind: #8154