@Scooletz 2025-02-03
@LukaszRozmej My take on it is that the global state provider should be updated only from a single thread. Having it updated from FCU handler may result in some funky behaviors. It' better to have FCU responsible for signaling where other components (`BlockProcessor`) use the provided information when needed