@LukaszRozmej 2025-02-03
> @LukaszRozmej My take on it is that the global state provider should be updated only from a single thread. Having it updated from FCU handler may result in some funky behaviors. It' better to have FCU responsible for signaling where other components (`BlockProcessor`) use the provided information when needed

Good point, but [NewPayload and FCU are always under single-threaded lock](https://github.com/NethermindEth/nethermind/blob/dfdd4161d05495c35ff45221b058f250595550a3/src/Nethermind/Nethermind.Merge.Plugin/EngineRpcModule.Paris.cs#L24).