@asdacap 2025-02-03
I _think_ the main purpose of the checkpoint is to revert when an exception happens. So my confusion is, why the `DoNotUpdateHead` do anything here. In any case, even with exception, it always reset the state at start block proccessing anyway, so why should it care to restore it?