@LukaszRozmej 2025-02-03
> I _think_ the main purpose of the checkpoint is to revert when an exception happens. So my confusion is, why the `DoNotUpdateHead` do anything here. In any case, even with exception, it always reset the state at start block proccessing anyway, so why should it care to restore it?

Because it is FCU that moves head state and not NewPayload? We probably could move state root on NewPayload if we wanted?