@LukaszRozmej 2025-02-04
> > > @asdacap
> > > > I think the main purpose of the checkpoint is to revert when an exception happens. So my confusion is, why the `DoNotUpdateHead` do anything here. In any case, even with exception, it always reset the state at start block processing anyway, so why should it care to restore it?
> > > 
> > > 
> > > First. This is not the change introduced in this PR, is it? Second, my understanding is that for options that do not contain `DoNotUpdateHead`, it should progress and do not reset the state. That was the behavior of the original.
> > 
> > 
> > I'm not saying that this PR introduce that change. I'm saying that if that logic is removed, that might solve your problem. The `DoNotUpdateHead` is also used to update the `BlockTree`'s Head which make much more sense than here.
> 
> I did the same with verkle, as it also had a similar problem with flat db layout.

Can you be more clear?