@Scooletz 2025-02-04
As far as I understand, we could ditch the final reset as @asdacap is proposing. Then, if a branch is properly processed, we leave the root as is. If it's not we reset to what has been captured before execution. The crux is that we always reset the root of the branch if it's different from the one that branch should be executed on. Having this said, removing of the reset to the checkpoint after a proper execution, seems like a good candidate for ensuring the checkpoint progression?